♻️ refactor(monitoring): extract codex settings modal
Move the Codex inspection settings modal into a focused presentation component.

Keep validation and persistence handlers in the page controller to preserve behavior.

Validated with type-check, focused monitoring tests, lint, and Vite production build.
